split up world code into public and private functions
